separation of concerns design pattern where the config managment is separate from repo_to_text and repo_to_c4 implementation
