+++ Salon/src/repo_to_text.py
@@ -21,7 +21,6 @@ Options:
 import argparse
 import os
 import sys
-import yaml
 import nltk
 from pathlib import Path
 from typing import Dict, Any, Set
@@ -29,122 +28,25 @@ from typing import Dict, Any, Set
 # Local modules
 from .directory_walker import add_visitor, walk_directory
 from .visitor_factory import get_visitors_for_text
+from .core.config_manager import ConfigManager  # Import ConfigManager
 
 nltk.download('punkt', quiet=True)
 
-def load_yaml(fname: str) -> Dict[str, Any]:
-    """
-    Load configuration from the YAML config file.
-    """
-    if not fname:
-        return {}
-    try:
-        with open(fname, 'r') as config_file:
-            return yaml.safe_load(config_file)
-    except FileNotFoundError:
-        print(f"Error: Configuration file '{fname}' not found.")
-        return {}
-    except yaml.YAMLError as e:
-        print(f"Error parsing YAML file: {e}")
-        return {}
-
-def parse_arguments() -> argparse.Namespace:
-    """
-    Parse command-line arguments.
-    """
-    parser = argparse.ArgumentParser(
-        description='Process a GitHub repository and concatenate file contents with optional readme summaries.',
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter
-    )
-
-    parser.add_argument(
-        '--cfg',
-        type=str,
-        default="config.yaml",
-        help='Path to the config.yaml file'
-    )
-
-    parser.add_argument(
-        '--repo_path',
-        type=str,
-        default='.',
-        help='Path to the local GitHub repository (absolute or relative)'
-    )
-
-    parser.add_argument(
-        '-w', '--max_words',
-        type=int,
-        default=200000,
-        help='Maximum number of words per output file'
-    )
-
-    parser.add_argument(
-        '-o', '--output_dir',
-        type=str,
-        default='.',
-        help='Directory to save the output files'
-    )
-
-    parser.add_argument(
-        '--skip_patterns',
-        type=str,
-        nargs='+',
-        help='Additional file patterns to skip (e.g., "*.md" "*.txt")'
-    )
-
-    parser.add_argument(
-        '--skip_dirs',
-        type=str,
-        nargs='+',
-        help='Additional directories to skip'
-    )
-
-    parser.add_argument(
-        '--model_type',
-        type=str,
-        default='braid_api',
-        help='Which summarisation model to use: "braid_api" or "local_gemini"'
-    )
-
-    parser.add_argument(
-        '-v', '--verbose',
-        action='store_true',
-        help='Enable verbose output'
-    )
-
-    return parser.parse_args()
-
-def validate_args(args: argparse.Namespace) -> None:
-    """
-    Validate and normalize arguments.
-    """
-    repo_path = Path(args.repo_path).resolve()
-    if not repo_path.exists():
-        raise ValueError(f"Repository path does not exist: {repo_path}")
-    if not repo_path.is_dir():
-        raise ValueError(f"Repository path is not a directory: {repo_path}")
-
-    output_dir = Path(args.output_dir).resolve()
-    output_dir.mkdir(parents=True, exist_ok=True)
-
-    args.repo_path = repo_path
-    args.output_dir = output_dir
 
 def main() -> int:
     """
     Main entry point for the script.
     """
-    args = parse_arguments()
-
+    config_manager = ConfigManager('Process a local GitHub repository')
     try:
-        validate_args(args)
+        config_manager.load_config()
+        args = config_manager.get_args()
+        config = config_manager.get_config()
     except ValueError as e:
         print(f"Error: {e}")
         return 1
 
     # Load config
-    config = load_yaml(args.cfg)
-
     skip_dirs: Set[str] = set(config.get("skip_dirs", []))
     if args.skip_dirs:
         skip_dirs.update(args.skip_dirs)
